Create LinearSearch.py
A searching algorithm that works in linear time. Given a list of items, traverses the list and returns the position of a specified search item. 

The un-Ordered Linear Search works with data sets not sorted. 

For efficiency, the ordered linear search searches the elements of a sorted list by comparing the current element in a given iteration to the search value. For list sorted in ascending order, if the current item is greater than the search value, then we know that the item is not in the list and vice versa.